Rename to genai-perf
