+++ src/c++/perf_analyzer/genai-perf/genai_perf/.gitignore
