+++ src/c++/perf_analyzer/genai-perf/genai_perf/utils.py
