+++ src/c++/perf_analyzer/genai-perf/pyproject.toml
@@ -25,7 +25,7 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 [project]
-name = "genai-pa"
+name = "genai-perf"
 readme = "README.md"
 description = "GenAI Perf Analyzer CLI - CLI tool to simplify profiling LLMs and Generative AI models with Perf Analyzer"
 dynamic = ["version"]
@@ -54,7 +54,7 @@ dependencies = [
 
 # CLI Entrypoint
 [project.scripts]
-genai-pa = "genai_pa.main:main"
+genai-perf = "genai_perf.main:main"
 
 [project.urls]
 "Homepage" = "https://github.com/triton-inference-server/"
@@ -66,7 +66,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.version]
-path = "genai_pa/__init__.py"
+path = "genai_perf/__init__.py"
 
 # Pre-commit hook tool configs
 [tool.codespell]
